requirements.txt
pycryptodome
python-dotenv
tqdm